Updated Duration Picker during Recipe Editing
