+++ Nextcloud Cookbook iOS Client/Views/Recipes/RecipeViewSections/RecipeKeywordSection.swift
@@ -1,5 +1,5 @@
 //
-//  KeywordPickerView.swift
+//  RecipeKeywordSection.swift
 //  Nextcloud Cookbook iOS Client
 //
 //  Created by Vincent Meilinger on 03.10.23.
@@ -8,7 +8,32 @@
 import Foundation
 import SwiftUI
 
+// MARK: - RecipeView Keyword Section
 
+struct RecipeKeywordSection: View {
+    @ObservedObject var viewModel: RecipeView.ViewModel
+    let columns: [GridItem] = [ GridItem(.flexible(minimum: 50, maximum: 200), spacing: 5) ]
+    
+    var body: some View {
+        CollapsibleView(titleColor: .secondary, isCollapsed: !UserSettings.shared.expandKeywordSection) {
+            Group {
+                if !viewModel.observableRecipeDetail.keywords.isEmpty && !viewModel.editMode {
+                    RecipeListSection(list: viewModel.observableRecipeDetail.keywords)
+                } else {
+                    Text(LocalizedStringKey("No keywords."))
+                }
+            }
+        } title: {
+            HStack {
+                SecondaryLabel(text: LocalizedStringKey("Keywords"))
+                Spacer()
+            }
+        }
+        .padding()
+    }
+}
+
+// MARK: - RecipeView Keyword Sheet View
 
 struct KeywordPickerView: View {
     @Environment(\.presentationMode) var presentationMode
@@ -31,7 +56,7 @@ struct KeywordPickerView: View {
             }
             TextField(title, text: $searchText)
                 .textFieldStyle(.roundedBorder)
-                .padding()
+                
             ScrollView {
                 LazyVGrid(columns: columns, spacing: 5) {
                     if searchText != "" {
@@ -94,7 +119,7 @@ struct KeywordPickerView: View {
             }
         }
         .navigationTitle(title)
-        .padding(5)
+        .padding()
         
     }
     
@@ -131,10 +156,36 @@ struct KeywordItemView: View {
         .padding()
         .background(
             RoundedRectangle(cornerRadius: 15)
-                .foregroundStyle(Color("backgroundHighlight"))
+                .foregroundStyle(.tertiary)
         )
         .onTapGesture {
             tapped(keyword)
         }
     }
 }
+
+
+// MARK: - Previews
+
+struct KeywordPickerView_Previews: PreviewProvider {
+    // Sample keywords for preview
+    static var sampleKeywords = [
+        RecipeKeyword(name: "Vegan", recipe_count: 10),
+        RecipeKeyword(name: "Meat", recipe_count: 5),
+        RecipeKeyword(name: "Gluten-Free", recipe_count: 8),
+        RecipeKeyword(name: "Difficult", recipe_count: 7),
+        RecipeKeyword(name: "Chinese", recipe_count: 3),
+        RecipeKeyword(name: "European", recipe_count: 5),
+        RecipeKeyword(name: "Easy", recipe_count: 1)
+  
+    ]
+    
+    // Sample selection for preview
+    @State static var selection: [String] = ["Vegan"]
+    
+    static var previews: some View {
+        KeywordPickerView(title: "Select Keywords", searchSuggestions: sampleKeywords, selection: $selection)
+            
+    }
+}
+
